move functions to their own files
